requirements.txt
google-generativeai
python-dotenv
cloudscraper
beautifulsoup4
requests